运行相关： 1、弹框提示未激活脚本 2、确认进行激活操作 3、取消激活操作 4、弹窗间隔1秒 5、启动appViewModel时先进行更新用户信息赋值相关过期时间等

+++ app/src/main/java/com/smart/autodaily/data/entity/UserInfo.kt
@@ -18,7 +18,7 @@ data class UserInfo (
     //类型名称
     @ColumnInfo(name = "key_type_name")   @SerializedName("key_type_name")   var keyTypeName: String?,
     //可激活个数
-    @ColumnInfo(name = "can_activate_num")   @SerializedName("can_activate_num")   var canActivateNum: Byte?,
+    @ColumnInfo(name = "can_activate_num")   @SerializedName("can_activate_num")   var canActivateNum: Int?,
     //过期时间
     @ColumnInfo(name = "expiration_time")   @SerializedName("expiration_time")   var expirationTime: String?,
     //注册时间
